feat: refactor test to kotest style
